# fix: Add ability for plainify to remove <style> and <script>