Added StripTag helper function to allow the ability to remove any tag and all
of its contents from a string.  Leveraged that new function within the plainify
function to remove <style> and <script> tags and contents.

Fixes #3235