@bep 2020-10-02
Hey, and thanks for this. The code looks clean and good, but I have one general comment:

* This issue is, in my head, relevant to all the callers of `StripHTML`, so this new logic might as well be included in that function.
* Also, the way `Plainify` works with this commit, you will get a fair amount of wasteful memory allocation, but moving this into `StripHTML` should give you some benchmarks to compare to.