@joshgerdes 2020-10-02
Heya 👋 , thanks for the feedback.

I agree it probably makes sense that someone using `StripHTML` would want those tag sections removed as well.  So were you thinking of just moving something like this into `StripHTML` function after the `stripHTMLReplacer` call:

```go
for _, value := range []string{"script", "style"} {
	stripTag(s, value)
}
```

Or moving the whole logic in and eliminating the `stripTag` function?

Or trying to integrate the logic of removing content with the tags in the current switch statement of `StripHTML`?