@bep 2020-10-02
>Or moving the whole logic in and eliminating the stripTag function?

See my comment about memory allocation. `StripHTML` should have a benchmark. If you run that before/after and compare with `benchstat` (also see my util project `gobench`), I suspect that with the `stripTag` function you will see a major hit in memory allocation. This will cost _something_, of course, but since we're running this on thousands of content pages, this matters.