@joshgerdes 2020-10-02
Thanks for the clarification and sorry I'm new to Go so still getting a handle on tooling, how allocation works, etc. You were correct though, here are the benchmark results I got comparing the two options.

With `stripTag` referenced in `StripHTML`:
```
name         old time/op    new time/op    delta
StripHTML-8    3.45µs ± 1%    4.84µs ± 1%   +40.17%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
StripHTML-8      896B ± 0%     2304B ± 0%  +157.14%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
StripHTML-8      4.00 ± 0%      8.00 ± 0%  +100.00%  (p=0.008 n=5+5)
```
With logic in `StripHTML`:
```
name         old time/op    new time/op    delta
StripHTML-8    3.45µs ± 1%    4.04µs ± 1%   +17.03%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
StripHTML-8      896B ± 0%     2080B ± 0%  +132.14%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
StripHTML-8      4.00 ± 0%      8.00 ± 0%  +100.00%  (p=0.008 n=5+5)
```

I have updated the PR, let me know if the latest is more what you were thinking.