@joshgerdes 2020-10-05
@bep is this what you were thinking?