@D4D3VD4V3 2020-12-18
I'd love to see this pushed upstream so I too can get more accurate estimates of words in a post.