@bep 2021-01-08
No, keep it open, it will be used, it fixes an important issue. I just need to understand the performance tradeoffs. I will look at this later.